Modify SaveButton
