+++ src/containers/group/create/GroupCreateName.js
@@ -1,9 +1,14 @@
 import React, {useState} from 'react';
-import {View, StyleSheet, ScrollView} from 'react-native';
+import {
+    View,
+    StyleSheet,
+    ScrollView,
+    TouchableOpacity,
+    Text,
+} from 'react-native';
 
 import TextInfo from '../../common/TextInfo';
 import GroupTextInput from './components/GroupTextInput';
-import SaveButton from '../../common/SaveButton';
 
 import colors from '../../../theme/colors';
 
@@ -15,14 +20,24 @@ function GroupCreateName({navigation}) {
         <View style={styles.container}>
             <ScrollView>
                 <TextInfo
-                title="그룹의 이름은 무엇인가요?"
-                subtitle="나만의 그룹 이름을 입력해보세요!"></TextInfo>
+                    title="그룹의 이름은 무엇인가요?"
+                    subtitle="나만의 그룹 이름을 입력해보세요!"></TextInfo>
                 <GroupTextInput
-                setButtonDisable={setGroupCreateButtonDisable}></GroupTextInput>
-                <SaveButton
-                text="다음으로"
-                buttonDisabled={groupCreateButtonDisable}
-                navigation={navigation}></SaveButton>
+                    setButtonDisable={
+                        setGroupCreateButtonDisable
+                    }></GroupTextInput>
+                {/* <SaveButton
+                    text="다음으로"
+                    buttonDisabled={groupCreateButtonDisable}
+                    navigation={navigation}></SaveButton> */}
+                <View style={{alignItems: 'center'}}>
+                    <TouchableOpacity
+                        onPress={() => {
+                            navigation.navigate('GroupCreateImage');
+                        }}>
+                        <Text>저장하기</Text>
+                    </TouchableOpacity>
+                </View>
             </ScrollView>
         </View>
     );
